Using lease for node-count in telemetry
